Rename main_window module to just main
