+++ alignment-software/alignment_software/__main__.py
@@ -1,4 +1,4 @@
-from .gui.window_main import MainWindow
+from .gui.main import MainWindow
 
 
 def main():
